+++ lib/rettulf.dart
@@ -5,12 +5,15 @@ export 'icon.dart';
 export 'string.dart';
 export 'widget_list/flex.dart';
 export 'widget_list/listview.dart';
+export 'widget_list/stack.dart';
 
 export 'widget/button.dart';
 export 'widget/center.dart';
 export 'widget/material.dart';
 export 'widget/padding.dart';
 export 'widget/sizedbox.dart';
+export 'widget/gesture.dart';
+export 'widget/safearea.dart';
 
 export 'buildcontext/general.dart';
 export 'buildcontext/show.dart';