+++ lib/widget/button.dart
@@ -1,4 +1,6 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:rettulf/widget/gesture.dart';
 
 extension WidgetButtonEx on Widget {
   TextButton textButton({
@@ -24,6 +26,54 @@ extension WidgetButtonEx on Widget {
           autofocus: autofocus,
           clipBehavior: clipBehavior,
           statesController: statesController,
+          child: this);
+
+  ElevatedButton elevatedButton({
+    Key? key,
+    VoidCallback? onPressed,
+    VoidCallback? onLongPress,
+    ValueChanged<bool>? onHover,
+    ValueChanged<bool>? onFocusChange,
+    ButtonStyle? style,
+    FocusNode? focusNode,
+    bool autofocus = false,
+    Clip clipBehavior = Clip.none,
+    MaterialStatesController? statesController,
+  }) =>
+      ElevatedButton(
+          key: key,
+          onPressed: onPressed,
+          onLongPress: onLongPress,
+          onHover: onHover,
+          onFocusChange: onFocusChange,
+          style: style,
+          focusNode: focusNode,
+          autofocus: autofocus,
+          clipBehavior: clipBehavior,
+          statesController: statesController,
+          child: this);
+
+  CupertinoButton cupertinoButton({
+    Key? key,
+    EdgeInsetsGeometry? padding,
+    Color? color,
+    Color disabledColor = CupertinoColors.quaternarySystemFill,
+    double minSize = kMinInteractiveDimensionCupertino,
+    double? pressedOpacity = 0.4,
+    BorderRadius? borderRadius = const BorderRadius.all(Radius.circular(8.0)),
+    Alignment alignment = Alignment.center,
+    VoidCallback? onPressed,
+  }) =>
+      CupertinoButton(
+          key: key,
+          onPressed: onPressed,
+          padding: padding,
+          color: color,
+          disabledColor: disabledColor,
+          minSize: minSize,
+          pressedOpacity: pressedOpacity,
+          borderRadius: borderRadius,
+          alignment: alignment,
           child: this);
 
   IconButton iconButton({