+++ lib/widget_list/stack.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/widgets.dart';
+
+extension StackEx on List<Widget> {
+  Stack stack({
+    Key? key,
+    AlignmentDirectional align = AlignmentDirectional.topStart,
+    TextDirection? textDirection,
+    StackFit fit = StackFit.loose,
+    Clip clip = Clip.hardEdge,
+  }) =>
+      Stack(
+        key: key,
+        alignment: align,
+        textDirection: textDirection,
+        fit: fit,
+        clipBehavior: clip,
+        children: this,
+      );
+}